requirements.txt
paho-mqtt
proxmoxer
requests
yaml
cryptography